% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg-Hght-Funct.R
\name{Avg_height_diff}
\alias{Avg_height_diff}
\title{Average Height Difference Function}
\usage{
Avg_height_diff(height, male = TRUE)
}
\arguments{
\item{height}{Height of person (in cm). I named this paramater as height because that is name the input.}

\item{male}{Variable to indicate the sex of the inputted height. I named this paramater as male because that corresponds to the input.}
}
\value{
A message (class type: Character) which contains the absolute value of the difference of the average height and the inputted height (in cm)
}
\description{
The \code{Avg_height_diff} function calculates the absolute difference of the average adult height per sex (178.1 cm for Canadian men & 164.7 cm for Canadian women) to an inputted height (in cm).
}
\examples{
Avg_height_diff(181, male = TRUE)
Avg_height_diff(114, male = FALSE)
}
